# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * eSIM reseller platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Package {
  packageCode: string;
  name: string;
  locationCode: string;
  locationName: string;
  dataGb: number;
  durationDays: number;
  priceUsd: number;
  retailPriceUsd: number;
  flagEmoji: string;
  valueScore?: number;
}

export interface Country {
  locationCode: string;
  locationName: string;
  flagEmoji: string;
  packageCount: number;
  minPriceUsd: number;
}

export interface OrderInput {
  packageCode: string;
  /** @nullable */
  familyMemberId: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  active: 'active',
  expired: 'expired',
  error: 'error',
} as const;

export interface Order {
  id: string;
  orderNo: string;
  packageCode: string;
  packageName: string;
  locationCode: string;
  locationName: string;
  dataGb: number;
  durationDays: number;
  priceUsd: number;
  status: OrderStatus;
  /** @nullable */
  qrCodeUrl?: string | null;
  /** @nullable */
  shortUrl?: string | null;
  /** @nullable */
  iccid?: string | null;
  /** @nullable */
  ac?: string | null;
  /** @nullable */
  activateTime?: string | null;
  /** @nullable */
  expiredTime?: string | null;
  /** @nullable */
  familyMemberId?: string | null;
  /** @nullable */
  familyMemberName?: string | null;
  createdAt: string;
}

export interface TopupInput {
  iccid: string;
  packageCode: string;
}

export type ManageEsimInputAction = typeof ManageEsimInputAction[keyof typeof ManageEsimInputAction];


export const ManageEsimInputAction = {
  suspend: 'suspend',
  unsuspend: 'unsuspend',
  revoke: 'revoke',
} as const;

export interface ManageEsimInput {
  iccid: string;
  action: ManageEsimInputAction;
}

export interface EsimSummary {
  totalActive: number;
  totalPending: number;
  totalExpired: number;
  totalSpentUsd: number;
  recentCountries?: string[];
}

export type WalletLevel = typeof WalletLevel[keyof typeof WalletLevel];


export const WalletLevel = {
  bronze: 'bronze',
  silver: 'silver',
  gold: 'gold',
  platinum: 'platinum',
} as const;

export interface Wallet {
  balanceUsd: number;
  level: WalletLevel;
  levelName: string;
  totalSpentUsd: number;
  discountPercent: number;
  /** @nullable */
  nextLevelSpendRequired?: number | null;
  levelupBonusUsd?: number;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  topup: 'topup',
  purchase: 'purchase',
  refund: 'refund',
  bonus: 'bonus',
  redeem: 'redeem',
} as const;

export interface Transaction {
  id: string;
  type: TransactionType;
  amountUsd: number;
  description: string;
  createdAt: string;
}

export type TopupOptionsOptionsItem = {
  amountUsd: number;
  label: string;
};

export interface TopupOptions {
  options: TopupOptionsOptionsItem[];
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface UserProfile {
  uid: string;
  email: string;
  displayName: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  travelFrequency: string | null;
  role?: UserProfileRole;
  referralCode: string;
  createdAt: string;
}

export interface ProfileUpdate {
  displayName?: string;
  travelFrequency?: string;
}

export interface Badge {
  id: string;
  name: string;
  description: string;
  earned: boolean;
  /** @nullable */
  earnedAt: string | null;
}

export interface ActiveEsimInfo {
  orderNo?: string;
  locationName?: string;
  dataGb?: number;
  expiredTime?: string;
  status?: string;
}

export interface FamilyMember {
  id: string;
  name: string;
  age: number;
  avatar: string;
  activeEsim?: ActiveEsimInfo;
  createdAt: string;
}

export interface FamilyMemberInput {
  name: string;
  age: number;
  avatar: string;
}

export interface FamilyMemberUpdate {
  name?: string;
  age?: number;
  avatar?: string;
}

export type ReferralInfoReferralsItem = {
  email: string;
  joinedAt: string;
  bonusEarnedUsd: number;
};

export interface ReferralInfo {
  referralCode: string;
  totalReferrals: number;
  totalBonusEarnedUsd: number;
  referrals?: ReferralInfoReferralsItem[];
}

export interface ReferralApplyInput {
  referralCode: string;
}

export interface RedeemInput {
  code: string;
}

export interface GenerateCodeInput {
  amountUsd: number;
}

export type RedeemCodeStatus = typeof RedeemCodeStatus[keyof typeof RedeemCodeStatus];


export const RedeemCodeStatus = {
  active: 'active',
  used: 'used',
} as const;

export interface RedeemCode {
  code: string;
  amountUsd: number;
  status: RedeemCodeStatus;
  /** @nullable */
  usedBy?: string | null;
  /** @nullable */
  usedAt?: string | null;
  createdAt: string;
}

export type TicketStatus = typeof TicketStatus[keyof typeof TicketStatus];


export const TicketStatus = {
  pending: 'pending',
  resolved: 'resolved',
} as const;

export type TicketRepliesItem = {
  message: string;
  isAdmin: boolean;
  createdAt: string;
};

export interface Ticket {
  id: string;
  subject: string;
  message: string;
  status: TicketStatus;
  replies?: TicketRepliesItem[];
  createdAt: string;
}

export interface TicketInput {
  subject: string;
  message: string;
}

export interface TicketReplyInput {
  message: string;
}

export type AdminStatsRevenueByDayItem = {
  date: string;
  revenueUsd: number;
};

export interface AdminStats {
  totalUsers: number;
  totalOrders: number;
  totalRevenueUsd: number;
  activeEsims: number;
  pendingTickets: number;
  revenueByDay?: AdminStatsRevenueByDayItem[];
}

export interface AdminUser {
  uid: string;
  email: string;
  displayName: string;
  balanceUsd: number;
  level: string;
  totalOrders: number;
  role: string;
  createdAt: string;
}

export interface BalanceAdjustInput {
  amountUsd: number;
  reason: string;
}

export interface SiteConfig {
  siteName: string;
  markupPercent: number;
  logoUrl: string;
  primaryColor?: string;
}

export interface SiteConfigUpdate {
  siteName?: string;
  markupPercent?: number;
  logoUrl?: string;
  primaryColor?: string;
}

export interface EsimAccessBalance {
  balanceUsd: number;
}

export interface GenericSuccess {
  success: boolean;
  message?: string;
}

export interface ApiError {
  error: string;
}

export type ListPackagesParams = {
/**
 * Country code filter (empty for all)
 */
locationCode?: string;
/**
 * Duration filter in days
 */
duration?: ListPackagesDuration;
sort?: ListPackagesSort;
};

export type ListPackagesDuration = typeof ListPackagesDuration[keyof typeof ListPackagesDuration];


export const ListPackagesDuration = {
  NUMBER_7: '7',
  NUMBER_15: '15',
  NUMBER_30: '30',
} as const;

export type ListPackagesSort = typeof ListPackagesSort[keyof typeof ListPackagesSort];


export const ListPackagesSort = {
  price_asc: 'price_asc',
  gb_asc: 'gb_asc',
  value_desc: 'value_desc',
} as const;

export type ListOrdersParams = {
status?: ListOrdersStatus;
familyMemberId?: string;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  pending: 'pending',
  active: 'active',
  expired: 'expired',
  all: 'all',
} as const;

export type ListUsersParams = {
search?: string;
};