# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: eSIM reseller platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /esim/order:
    post:
      operationId: createOrder
      tags: [esim]
      summary: Purchase an eSIM package using wallet balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Insufficient balance or invalid package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /esim/orders:
    get:
      operationId: listOrders
      tags: [esim]
      summary: List user's eSIM orders
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, active, expired, all]
        - in: query
          name: familyMemberId
          schema:
            type: string
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /esim/orders/{orderNo}:
    get:
      operationId: getOrder
      tags: [esim]
      summary: Get order details with QR code
      parameters:
        - in: path
          name: orderNo
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /esim/topup:
    post:
      operationId: topupEsim
      tags: [esim]
      summary: Top up an existing eSIM with a data package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupInput"
      responses:
        "200":
          description: Topup successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /esim/manage:
    post:
      operationId: manageEsim
      tags: [esim]
      summary: Suspend, unsuspend, or revoke an eSIM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManageEsimInput"
      responses:
        "200":
          description: Action successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericSuccess"
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin only)
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{uid}/balance:
    patch:
      operationId: adjustUserBalance
      tags: [admin]
      summary: Adjust a user's wallet balance (admin only)
      parameters:
        - in: path
          name: uid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BalanceAdjustInput"
      responses:
        "200":
          description: Balance adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericSuccess"
  /admin/config:
    get:
      operationId: getAdminConfig
      tags: [admin]
      summary: Get site configuration (admin only)
      responses:
        "200":
          description: Config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteConfig"
    patch:
      operationId: updateAdminConfig
      tags: [admin]
      summary: Update site configuration (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteConfigUpdate"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteConfig"
...
components:
  schemas:
...
    OrderInput:
      type: object
      required: [packageCode, familyMemberId]
      properties:
        packageCode:
          type: string
        familyMemberId:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, orderNo, packageCode, packageName, locationCode, locationName, dataGb, durationDays, priceUsd, status, createdAt]
      properties:
        id:
          type: string
        orderNo:
          type: string
        packageCode:
          type: string
        packageName:
          type: string
        locationCode:
          type: string
        locationName:
          type: string
        dataGb:
          type: number
        durationDays:
          type: integer
        priceUsd:
          type: number
        status:
          type: string
          enum: [pending, active, expired, error]
        qrCodeUrl:
          type: ["string", "null"]
        shortUrl:
          type: ["string", "null"]
        iccid:
          type: ["string", "null"]
        ac:
          type: ["string", "null"]
        activateTime:
          type: ["string", "null"]
        expiredTime:
          type: ["string", "null"]
        familyMemberId:
          type: ["string", "null"]
        familyMemberName:
          type: ["string", "null"]
        createdAt:
          type: string
    TopupInput:
      type: object
      required: [iccid, packageCode]
      properties:
        iccid:
          type: string
        packageCode:
          type: string
    ManageEsimInput:
      type: object
      required: [iccid, action]
      properties:
        iccid:
          type: string
        action:
          type: string
          enum: [suspend, unsuspend, revoke]
...
    AdminUser:
      type: object
      required: [uid, email, displayName, balanceUsd, level, totalOrders, role, createdAt]
      properties:
        uid:
          type: string
        email:
          type: string
        displayName:
          type: string
        balanceUsd:
          type: number
        level:
          type: string
        totalOrders:
          type: integer
        role:
          type: string
        createdAt:
          type: string
    BalanceAdjustInput:
      type: object
      required: [amountUsd, reason]
      properties:
        amountUsd:
          type: number
        reason:
          type: string
    SiteConfig:
      type: object
      required: [siteName, markupPercent, logoUrl]
      properties:
        siteName:
          type: string
        markupPercent:
          type: integer
        logoUrl:
          type: string
        primaryColor:
          type: string
    SiteConfigUpdate:
      type: object
      properties:
        siteName:
          type: string
        markupPercent:
          type: integer
        logoUrl:
          type: string
        primaryColor:
          type: string
...
    GenericSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string