# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * eSIM reseller platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  orderNo: string;
  packageCode: string;
  packageName: string;
  locationCode: string;
  locationName: string;
  dataGb: number;
  durationDays: number;
  priceUsd: number;
  status: OrderStatus;
  /** @nullable */
  qrCodeUrl?: string | null;
  /** @nullable */
  shortUrl?: string | null;
  /** @nullable */
  iccid?: string | null;
  /** @nullable */
  ac?: string | null;
  /** @nullable */
  activateTime?: string | null;
  /** @nullable */
  expiredTime?: string | null;
  /** @nullable */
  familyMemberId?: string | null;
  /** @nullable */
  familyMemberName?: string | null;
  createdAt: string;
}